clippy: clean the whole workspace
Mechanical bulk via cargo clippy --fix --workspace, then by hand:

- hi-cli: clamp() in diff_lab; collapse identical if/else arms in eval
  and report status classification; box Response::Event envelope (wire
  enum was padded to SessionEvent size); #[allow(too_many_arguments)]
  with justification on race/report CLI-plumbing fns
- hi-policy: box Authorization::RequiresApproval (cold path; variant
  never constructed yet, so no call-site churn); allow on the
  field-per-field approval_request builder
- hi-shell: box AcpUiEvent::Update (440 B session update vs 8 B flush)
- hi-replay: rename InteractiveEventReplay::next to advance to avoid
  Iterator::next confusion (crate-internal; no external users)
- hi-control/hi-agent: type aliases for complex tuple types (ScopeRow,
  PendingCheck, BackendCacheEntry)
- hi-mcp: discover_servers takes &Path; drop now-unused PathBuf import
- hi-tools: hf_downloader open with explicit truncate(true); allow on
  the 1:1 range-download retry fn
- hi-agent-runtime: unused inspect_err binding
- hi-eval/hi-codebase-graph/hi-tui/hi-agent: allow(too_many_arguments)
  on fns whose params mirror CLI flags or manifest fields 1:1

cargo clippy --workspace: zero lints; hi 251/251, hi-agent 916/916,
hi-tui 355/355, hi-eval 253/253, hi-replay/hi-control/hi-mcp/hi-shell/
hi-policy/hi-diff all green

+++ crates/hi-tools/src/hf.rs
@@ -980,7 +980,7 @@ pub fn available_space_bytes(path: &Path) -> Option<u64> {
         return None;
     }
     let stats = unsafe { stats.assume_init() };
-    Some((stats.f_bavail as u64).saturating_mul(stats.f_frsize as u64))
+    Some((stats.f_bavail as u64).saturating_mul(stats.f_frsize))
 }
 
 #[cfg(not(unix))]
